requirements.txt
requests
beautifulsoup4
openai
networkx
pyvis
python-dotenv
flask
duckduckgo-search
